+++ examples/data-sources/ciphertrust_gcp_connection/data_source.tf
@@ -1,0 +1,4 @@
+# Get the GCP connection data using the connection name
+data "ciphertrust_gcp_connection" "by_connection_name" {
+  name  = "connection-name"
+}