+++ examples/provider/provider.tf
@@ -1,5 +1,5 @@
 provider "ciphertrust" {
-  address           = "https://34.207.194.87"
+  address           = "https://ip_or_hostname_of_cm"
   username          = "username"
   password          = "password"
   domain            = "domain"