+++ examples/resources/ciphertrust_policies/resource.tf
@@ -1,0 +1,11 @@
+resource "ciphertrust_policies" "policy" {
+    name    =   "policyReadKeyOnly"
+    actions =   ["ReadKey"]
+    allow   =   true
+    effect  =   "allow"
+    conditions {
+        path   = "context.resource.alg"
+        op     = "equals"
+        values = ["aes","rsa"]
+    }
+}