+++ examples/resources/ciphertrust_policy_attachment/resource.tf
@@ -1,0 +1,8 @@
+resource "ciphertrust_policy_attachments" "policyattachment" {
+  policy = ciphertrust_policies.policy.id
+  principal_selector = <<-EOT
+    {
+      "groups" : ["admin"]
+    }
+  EOT
+}