+++ sample-scripts/policies/cloud-key-manager/google/allow-keyring-create/gcp_vars.tf
@@ -1,0 +1,14 @@
+variable "gcp_key_file" {
+  type    = string
+  default = "../../../server_certs/gcp-key-file.json"
+}
+
+variable "gcp_project" {
+  type    = string
+  default = "gemalto-kyloeng"
+}
+
+variable "keyring_ex1" {
+  type    = string
+  default = "projects/gemalto-kyloeng/locations/global/keyRings/CCKM-Automation1"
+}